cloud_services/futures_ingestion/requirements.txt
yfinance==0.2.28
pandas==2.1.4
google-cloud-bigquery==3.11.4
